refactor: remove createInfra function and integrate resource creation directly in sst.config
